requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
Pillow==8.2.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.2.0
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.7.0
